プレイヤーを狙うAIを作成。しかし、対エネミーWaveでもプレイヤーを狙う個体があったり、動きがぎこちなくなったりするようになった。 プレイヤーを武器で攻撃しても資源をドロップしないバグが発生。おそらく前のコミットが原因か。詳細を追って調査する。

+++ ScrapScramble/Assets/Scripts/Player/PlayerMovement.cs
@@ -284,7 +284,7 @@ public class PlayerMovement : MonoBehaviour
         }
 
         // 角度の差を算出
-        float angle = 360 / dropMass;
+        float angle = 360.0f / (float)dropMass;
 
         // 排出する
         for (int i = 0; i < dropMass; ++i)
